Exclude patches that are meant for the next version
Output modes support is a feature that will be accepted for
QtCompositor 5.6 and so we don't need this now.

The only known user of this API is Green Island which is compatible
with both API versions depending on which Qt version is used, hence
it will make use of this new API when it's built against Qt 5.6.